Fixed direct link to edit others posts

+++ Application/app/Http/Controllers/PostsController.php
@@ -86,7 +86,6 @@ class PostsController extends Controller
     {
 
         $post = Post::find($id);
-
         // check for correct user
         if(auth()->user()->id !== $post->user_id){
             return redirect('/posts')->with('error', 'Unauthorized Page');
@@ -124,6 +123,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+        
         // delete
         $post->delete();
         return redirect('/posts')->with('success', 'Post Delete');
